types/big.js: Don't export global 'Big' identifier
Only <script src=...> users will get the global definition.  Node.js and
ES module users won't, so TypeScript should not let you use 'Big' unless
you import the module.

Also modified tests to make sure the 'Big' symbol and the default export
can be used as types.
