+++ types/big.js/tsconfig.json
@@ -19,7 +19,8 @@
     "files": [
         "index.d.ts",
         "test/big.js-module-tests.ts",
-        "test/big.js-global-tests.ts",
-        "test/big.js-import-default-tests.ts"
+        "test/big.js-import-default-tests.ts",
+        "test/big.js-import-compatibility-tests.ts",
+        "test/big.js-global-module-tests.ts"
     ]
 }
